+++ bridgetree/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to Rust's notion of
 
 ## [Unreleased]
 
+### Removed
+
+- The `testing` module has been removed in favor of depending on
+  `incrementalmerkletree::testing`.
+
 ## [bridgetree-v0.2.0] - 2022-05-10
 
 The `bridgetree` crate is a fork of `incrementalmerkletree`, with the contents
@@ -84,7 +89,7 @@ The `Tree` interface reflects the renaming of `witness` to `mark` described abov
 
 ### Removed relative to `incrementalmerkletree-0.3.0`
 
-- `bridgetree::Leaf` 
+- `bridgetree::Leaf`
 - `bridgetree::AuthFragment`
 - `NonEmptyFrontier::size`
 - `NonEmptyFrontier::max_altitude`